requirements.txt
py-cord == 2.2.2
python-dotenv